SHL-Recommender-main/backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pandas
numpy
httpx
python-dotenv
